Moves components to dir
